Version 1.4. Released the date 21/06/2013.
Changes in the About page. Now it shows an Historical Log of changes in the Framework and the version on development.
Refactoring View library; private and protected functions and properties now uses underscore before property name.
Improved View; now it works with "Chunk" concepts, so multiple view chunks can be used. It allows inserting the Chunks in a specific order.
Modified View and Controller jurisdiction: JS and CSS libraries are added in the Controller, as the View should only allowed the controllers to use it.
Added Meta feature in the View: now the general Header will write the Metas specified in the View by the Controllers.
Modified Error page; a little Error CSS to show the improvements with the CSS and views.
Modified Index page; use of multiple views with a specific order to show improvements with Chunks.