Version 1.81, released 16/07/2013.
Fixed issues with Exceptions catching in the Bootstrap.
Fixed issues where the core Exception class was thrown instead of the engine one.
Fixed issues with the sample application.
Minor changes in the CSS of the Error page for the sample application.
Realized breach in the execute statement in Database Class, due to a bug in the UserModel. Make sure to print any exception from PDO to learn about how it might work in case of SQL issues.